refactor: remove default/native split
