+++ src/index.tsx
@@ -2,6 +2,6 @@ export { unstable_configureTextCompat } from './compat';
 export type { TextCompatConfig } from './compat';
 export { PlainText } from './PlainText';
 export type { PlainTextProps, PlainTextStyle } from './PlainText';
-export { unstable_mapPlainTextProps } from './PlainText.native';
+export { unstable_mapPlainTextProps } from './PlainText';
 export { default as unstable_NativePlainText } from './PlainTextViewNativeComponent';
 export type { NativeProps as unstable_NativePlainTextProps } from './PlainTextViewNativeComponent';
